refactor(engine): widen CompiledNode.muteBit to uint32 (ADR-0016 step 1)
Pure type widening ahead of the multi-word mute mask: muteBit and kNoMuteBit
go uint8 -> uint32 (absorbed into existing struct slack; the sizeof<=64 and
trivially-copyable static_asserts still hold). Storage stays a single uint64 and
the <64 guards are untouched, so behaviour is identical. Local: YesDawGraphCheck,
YesDawBuilderCheck, YesDawMixerMutePolicyCheck, YesDawMixerProjectionCheck green.

+++ src/engine/GraphBuilder.h
@@ -546,7 +546,7 @@ private:
             cn.outputSlot  = outputSlot;
             cn.pathLatency = item.pathLatency;
             cn.delayCacheKey = item.delayCacheKey;
-            cn.muteBit     = compiledIdx < 64u ? static_cast<std::uint8_t> (compiledIdx) : kNoMuteBit;
+            cn.muteBit     = compiledIdx < 64u ? static_cast<std::uint32_t> (compiledIdx) : kNoMuteBit;
             cn.kind        = item.kind;
             cn.aliasOk     = aliasOk;
